requirements.txt
python-telegram-bot[rate-limiter]==21.6
aiosqlite==0.20.0
APScheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
openai>=1.51.0
